Add better exception handling
All of the functions now return an exception instead of just panicking. `ReadFile()` now uses the absolute value of the passed in file path. All the arguments are now renamed to "filePath" instead of just as "fileName"; which implies that just the name of the file should be passed in instead of the complete path. All the doc comments are now updated.